Added render_upload.py
The script will generate a shell script which will upload all files
under the output/ folder to the ftp server.
This explicit step was necessary because the ftp server does not allow
file globbing only explicit target file naming.
The Python script will produce a ftp script of the form
    put local_src/site1.html server_dest/site1.html
    put local_src/site2.html server_dest/site2.html
    ...